Refactor tax report tests: extract reusable components into `TaxReportRowRepository`, `TaxReportRowMatcher`, and `TaxInfoLoadingCache` for improved readability and modularization.
